+++ src/server/app.js
@@ -11,6 +11,7 @@ var error = require('./error.js');
 var projects = require('./routes/projects');
 var admin = require('./routes/admin');
 var boards = require('./routes/boards');
+var courses = require('./routes/courses');
 var statusCodes = require('http-status-codes');
 
 debug.log = console.info.bind(console);
@@ -20,25 +21,27 @@ if (process.env.NODE_ENV !== 'production') app.use(logger('dev'));
 
 var apiv1 = express.Router();
 
+apiv1.use(bodyParser());
 apiv1.use(bodyParser.urlencoded({ extended: false }));
 apiv1.use(bodyParser.json());
 
 
-apiv1.use('/user', users.publicRoutes);
-
-apiv1.use('/remote', boards.remoteRoute);
+apiv1.use('/users', users.publicRoutes);
+apiv1.use('/remote', boards.remoteRoutes);
 
 apiv1.use(users.security);
 
-apiv1.use('/boards', boards.boardRoute);
-
-apiv1.use('/user', users.privateRoutes);
-
-apiv1.use('/projects', projects.projectsRouter);
+apiv1.use('/boards', boards.privateRoutes);
+apiv1.use('/users', users.privateRoutes);
+apiv1.use('/projects', projects.privateRoutes);
+apiv1.use('/courses', courses.privateRoutes);
 
 apiv1.use(admin.adminSecurity);
 
-apiv1.use('/admin', admin.adminRoute);
+apiv1.use('/users', users.adminRoutes);
+apiv1.use('/boards', boards.adminRoutes);
+apiv1.use('/courses', courses.adminRoutes);
+apiv1.use('/boards', boards.adminRoutes);
 
 app.use('/api/v1', apiv1);
 
@@ -53,12 +56,12 @@ app.use(function(err, req, res, next) {
 		if (err.status == statusCodes.INTERNAL_SERVER_ERROR) {
 			error.sendError(res, error.serverError('Something went wrong with your request. Try again later!'));
 			console.error(err);
+		} else {
+			error.sendError(res, err);
 		}
-		error.sendError(res, err);
 	} else {
 		error.sendError(res, error.notFound('Page not found'));
 	}
-	next;
 });
 
 module.exports = app;
