+++ src/server/database/database.js
@@ -57,9 +57,9 @@ try {
 
 var user = require('./user.js');
 var workspace = require('./workspace.js');
-var course = require('./courses.js');
+var course = require('./course.js');
 var board = require('./board.js');
-var image = require ('./raspberrypi.js');
+var image = require('./raspberrypi.js');
 module.exports.user = user;
 module.exports.workspace = workspace;
 module.exports.course = course;
