+++ src/server/routes/projects.js
@@ -4,11 +4,11 @@ var express = require('express');
 var debug = require('debug')('wyliodrin-lab-server:projects-routes');
 var db = require('../database/database.js');
 var error = require('../error.js');
-var projectApp = express.Router();
+var privateApp = express.Router();
 
 debug.log = console.info.bind(console);
 
-projectApp.post('/new_project', async(req, res, next) => {
+privateApp.post('/add', async(req, res, next) => {
 	var userId = req.user.userId;
 	var projectName = req.body.projectName;
 	var result = await db.workspace.createProject(userId, projectName);
@@ -21,7 +21,7 @@ projectApp.post('/new_project', async(req, res, next) => {
 	}
 });
 
-projectApp.post('/list_projects', async(req, res, next) => {
+privateApp.post('/list', async(req, res, next) => {
 	var userId = req.user.userId;
 	try {
 		var projects = await db.workspace.listProjects(userId);
@@ -33,4 +33,4 @@ projectApp.post('/list_projects', async(req, res, next) => {
 	res.status(200).send(projects);
 });
 
-module.exports.projectsRouter = projectApp;
+module.exports.privateRoutes = privateApp;
